refactored accouncement routines
