+++ src/providers/queries/index.ts
@@ -28,8 +28,6 @@ export const initQueryClient = () => {
           return query.queryKey[3] === 0; // only dehydrate first page of waves
         case QUERIES.NOTIFICATIONS.GET:
           return query.queryKey[2] === ''; // only dehydrate first page of notifications
-        case QUERIES.NOTIFICATIONS.GET_ANNOUNCEMENTS:
-          return false;
         default:
           return true;
       }
@@ -58,3 +56,4 @@ export * from './postQueries';
 export * from './walletQueries';
 export * from './leaderboardQueries';
 export * from './settingsQueries';
+export * from './announcementsQueries';
