+++ src/providers/queries/queryKeys.ts
@@ -7,7 +7,9 @@ const QUERIES = {
   },
   NOTIFICATIONS: {
     GET: 'QERUY_GET_NOTIFICATIONS',
-    GET_ANNOUNCEMENTS: 'QERUY_GET_ANNOUNCEMENTS'
+  },
+  ANNOUNCEMENTS:{
+    GET: 'QERUY_GET_ANNOUNCEMENTS'
   },
   SNIPPETS: {
     GET: 'QUERY_GET_SNIPPETS',
